pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memtotal"
version = "0.1.0"
description = "Bootstrap harness for MemTOTAL experiments."
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0",
  "datasets>=4.0.0",
  "rouge-score>=0.1.2",
  "transformers>=4.55.0",
  "accelerate>=1.10.0",
  "safetensors>=0.7.0",
  "sentencepiece>=0.2.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
